Simplified chat view controller
